+++ test/yog/property/tree_decomposition_test.exs
@@ -1,0 +1,5 @@
+defmodule Yog.Property.TreeDecompositionTest do
+  use ExUnit.Case
+  doctest Yog.Property.TreeDecomposition
+  alias Yog.Property.TreeDecomposition
+end